requirements.txt
autopep8==1.5.6
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colored==1.4.2
idna==2.10
pycodestyle==2.7.0
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
tzlocal==2.1
urllib3==1.26.4
